# api/openapi.yaml
openapi: 3.1.0
info:
  title: AtApp Gateway API
  description: API Gateway interface for client-service interactions in AtApp college attendance system.
  version: 1.0.0
servers:
  - url: /api/v1
paths:
...
  /platform/colleges:
    post:
      summary: Onboard a new college tenant
      description: >
        Platform-level endpoint with no tenant context yet; gated by the shared
        X-Platform-Admin-Key header instead of a per-tenant JWT. Creates the
        college and its first admin user atomically.
      security:
        - PlatformAdminKey: []
      parameters:
        - name: X-Platform-Admin-Key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - adminName
                - adminEmail
                - adminPassword
              properties:
                name:
                  type: string
                  example: MIT
                adminName:
                  type: string
                  example: Root Admin
                adminEmail:
                  type: string
                  format: email
                adminPassword:
                  type: string
                  format: password
                  minLength: 8
      responses:
        '201':
          description: College and initial admin created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  name:
                    type: string
                  adminUserId:
                    type: string
                    format: uuid
        '400':
          description: Missing or invalid fields
        '403':
          description: Missing or invalid platform admin key
...
  /teacher/sessions:
    post:
      summary: Start a class session
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - sectionId
              properties:
                sectionId:
                  type: string
                  format: uuid
                geofenceOverride:
                  type: object
                  properties:
                    lat:
                      type: number
                    lng:
                      type: number
                    radiusM:
                      type: number
      responses:
        '201':
          description: Session successfully started
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: string
                    format: uuid
                  currentCode:
                    type: string
                    example: "482910"
                  codeExpiresInSeconds:
                    type: integer
                    example: 10
        '403':
          description: Unauthorized role (requires teacher)
        '404':
          description: Section not found or not assigned to this teacher
        '409':
          description: A session is already active for this section
  /teacher/sessions/{sessionId}/code:
    get:
      summary: Get or rotate the current check-in code for an active session
      security:
        - BearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Current (or newly rotated) code
          content:
            application/json:
              schema:
                type: object
                properties:
                  currentCode:
                    type: string
                  codeExpiresInSeconds:
                    type: integer
        '404':
          description: Active session not found or not assigned to this teacher
  /teacher/sessions/{sessionId}/end:
    post:
      summary: End a class session
      security:
        - BearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Session successfully ended
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: string
                    format: uuid
                  summary:
                    type: object
                    properties:
                      presentCount:
                        type: integer
                      absentCount:
                        type: integer
                      overrideCount:
                        type: integer
        '404':
          description: Session not found or not assigned to this teacher
        '409':
          description: Session is already ended
  /student/checkin:
    post:
      summary: Submit a session check-in
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
              properties:
                code:
                  type: string
                  example: "482910"
                bssid:
                  type: string
                  example: "00:0a:95:9d:68:16"
                gps:
                  type: object
                  required:
                    - lat
                    - lng
                  properties:
                    lat:
                      type: number
                      example: 42.3601
                    lng:
                      type: number
                      example: -71.0942
      responses:
        '200':
          description: Checkin processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: string
                    enum: [accepted, rejected]
                    example: accepted
                  rejectionReason:
                    type: string
                    nullable: true
                    example: "Out of classroom geofence"
        '400':
          description: Invalid request parameters
        '404':
          description: No active class session found for the student's enrolled courses
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    PlatformAdminKey:
      type: apiKey
      in: header
      name: X-Platform-Admin-Key